Restructured express for `supertest` configuration
